requirements.txt
fastapi[standard]
sqlalchemy
pyjwt
docker
websockets 